UI Improvements & Code Cleanup
Removed unused imports, adjusted source file structure and solution output paths.

Details:

- Improved UI layout
- Cleaned up UI strings
- Added readme
- Added license
