+++ .gitignore
@@ -14,6 +14,7 @@
 *.userprefs
 
 # Build results
+Build
 [Dd]ebug/
 [Dd]ebugPublic/
 [Rr]elease/
@@ -29,6 +30,7 @@ bld/
 
 # Visual Studio 2015/2017 cache/options directory
 .vs/
+Source/obj
 # Uncomment if you have tasks that create the project's static files in wwwroot
 #wwwroot/
