--- SGDBMetadata/HeroModel.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SGDBMetadata
-{
-    public class HeroModel : MediaModel
-    {
-        public string style { get; set; }
-    }
-}
