+++ Source/GridModel.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SGDBMetadata
+﻿namespace SGDBMetadata
 {
     public class GridModel : MediaModel
     {
