+++ Source/HeroModel.cs
@@ -1,0 +1,7 @@
+﻿namespace SGDBMetadata
+{
+    public class HeroModel : MediaModel
+    {
+        public string style { get; set; }
+    }
+}
