+++ Source/Properties/AssemblyInfo.cs
