+++ Source/SGDBMetadataSettingsView.xaml.cs
@@ -1,17 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
+﻿using System.Collections.Generic;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 
 namespace SGDBMetadata
 {
@@ -23,23 +11,23 @@ namespace SGDBMetadata
             
             cmbDimensions.ItemsSource = new List<string>
             {
-                "any",
+                "Any",
                 "460x215",
                 "920x430",
                 "600x900",
                 "342x482",
-                "legacy"
+                "Legacy"
             };
             cmbDimensions.SelectedIndex = 0;
 
             cmbStyles.ItemsSource = new List<string>
             {
-                "any",
-                "alternate",
-                "blurred",
-                "white_logo",
-                "material",
-                "no_logo"
+                "Any",
+                "Alternate",
+                "Blurred",
+                "White Logo",
+                "Material",
+                "No Logo"
             };
             cmbStyles.SelectedIndex = 0;
         }
