+++ Source/SGDBService.cs
@@ -1,12 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Newtonsoft;
 using RestSharp;
 using RestSharp.Authenticators;
-using SGDBMetadata;
 using Playnite.SDK;
 
 namespace SGDBMetadata
