+++ Source/extension.yaml
@@ -1,4 +1,4 @@
-﻿Name: Steam Grid DB Metadata Provider
+﻿Name: SteamGridDB Metadata Provider
 Author: cooperate
 Version: 0.4
 Module: SGDBMetadata.dll
