+++ SteamGridDB Metadata.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29806.167
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SGDBMetadata", "SGDBMetadata\SGDBMetadata.csproj", "{13C224DB-2CF0-4876-A7FD-0DB4F4D6690F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SGDBMetadata", "Source\SteamGridDB Metadata.csproj", "{13C224DB-2CF0-4876-A7FD-0DB4F4D6690F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
